Allow install on Windows
Until ruby-3.3 syslog is shipped with ruby.
It is compiled when the POSIX API is available.
Otherwise it is skipped and a require 'syslog' fails at runtime.

But syslog will not be included in ruby-3.4 any longer.
So it's necessary to add it as a gem dependency.
Unfortunately this dependency can not be resolved on Windows, since the gem can't be installed there.

This patch allows syslog to be installed on Windows.
A require 'syslog' fails with a proper error message, if it's not rescued.
That way the app/gem can declare dependencies platform independent.
And it can use an alternative approach on platforms not supporting syslog.